feat: Add authentication check for all Trainer/Trainee functions except profile creation

+++ src/main/java/com/epam/gymcrm/api/controller/TraineeController.java
@@ -3,10 +3,14 @@ package com.epam.gymcrm.api.controller;
 import com.epam.gymcrm.api.payload.request.*;
 import com.epam.gymcrm.api.payload.response.*;
 import com.epam.gymcrm.domain.service.TraineeService;
+import com.epam.gymcrm.exception.UnauthorizedException;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import static com.epam.gymcrm.api.auth.AuthSessionManager.isLoggedInd;
+import static com.epam.gymcrm.api.auth.AuthSessionManager.logout;
 
 @RestController
 @RequestMapping(value = "/api/v1/trainees", produces = "application/json")
@@ -24,29 +28,64 @@ public class TraineeController {
     }
 
     @GetMapping("/profile")
-    public ResponseEntity<TraineeProfileResponse> getTraineeByUsername(
-            @RequestParam(name = "username") String username
-    ) {
-        //traineeService.isTraineeCredentialsValid(username, password);
-        return ResponseEntity.ok(traineeService.findByUsername(username));
+    public ResponseEntity<TraineeProfileResponse> getTraineeByUsername(@RequestParam(name = "username") String username) {
+        // Check if the user is authenticated
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
+        TraineeProfileResponse traineeProfileResponse = traineeService.findByUsername(username);
+
+        // Logout the user immediately after the successful change.
+        logout(username);
+
+        return ResponseEntity.ok(traineeProfileResponse);
     }
 
     @PutMapping
     public ResponseEntity<TraineeProfileUpdateResponse> updateTrainee(@RequestBody @Valid TraineeUpdateRequest traineeUpdateRequest) {
-        return new ResponseEntity<>(traineeService.update(traineeUpdateRequest), HttpStatus.OK);
+        String username = traineeUpdateRequest.username();
+
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
+        TraineeProfileUpdateResponse response = traineeService.update(traineeUpdateRequest);
+
+        logout(username);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @DeleteMapping("/{username}")
     public ResponseEntity<Void> deleteTraineeByUsername(@PathVariable(name = "username") String username) {
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
         traineeService.deleteTraineeByUsername(username);
+
+        logout(username);
+
         return ResponseEntity.ok().build();
     }
 
     @PutMapping("/trainers")
-    public ResponseEntity<TraineeTrainerUpdateResponse> updateTraineeTrainers(
-            @RequestBody @Valid TraineeTrainerUpdateRequest request
-    ) {
-        return ResponseEntity.ok(traineeService.updateTraineeTrainers(request));
+    public ResponseEntity<TraineeTrainerUpdateResponse> updateTraineeTrainers(@RequestBody @Valid TraineeTrainerUpdateRequest request) {
+        String username = request.traineeUsername();
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
+        TraineeTrainerUpdateResponse response = traineeService.updateTraineeTrainers(request);
+
+        logout(username);
+
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping("/trainings")
@@ -57,22 +96,48 @@ public class TraineeController {
             @RequestParam(name = "trainerName", required = false) String trainerName,
             @RequestParam(name = "trainingType", required = false) String trainingType
     ) {
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
         TraineeTrainingsFilter filter = new TraineeTrainingsFilter(
                 username, periodFrom, periodTo, trainerName, trainingType
         );
-        return ResponseEntity.ok(traineeService.getTraineeTrainings(filter));
+        TraineeTrainingsListResponse response = traineeService.getTraineeTrainings(filter);
+
+        logout(username);
+
+        return ResponseEntity.ok(response);
     }
 
     @PatchMapping("/status")
     public ResponseEntity<Void> activateTrainee(@RequestBody @Valid UpdateActiveStatusRequest updateActiveStatusRequest) {
+        String username = updateActiveStatusRequest.username();
+
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
         traineeService.updateActivateStatus(updateActiveStatusRequest);
+
+        logout(username);
+
         return ResponseEntity.ok().build();
     }
 
     @GetMapping("/unassigned-trainers")
-    public ResponseEntity<UnassignedActiveTrainerListResponse> getUnassignedTrainersForTrainee(
-            @RequestParam(name = "username") String username
-    ) {
-        return ResponseEntity.ok(traineeService.getUnassignedActiveTrainersForTrainee(username));
+    public ResponseEntity<UnassignedActiveTrainerListResponse> getUnassignedTrainersForTrainee(@RequestParam(name = "username") String username) {
+        if (!isLoggedInd(username)) {
+            // If not logged in, throw an error.
+            throw new UnauthorizedException("Trainee not authenticated, please login first!");
+        }
+
+        UnassignedActiveTrainerListResponse response = traineeService.getUnassignedActiveTrainersForTrainee(username);
+
+        logout(username);
+
+        return ResponseEntity.ok(response);
     }
 }
